refactor(typage): les loaders cessent de re-typer — suppression des re-CAST (#293)
ADR-0035 : le loader ne connaît plus aucun type, il trust ce que dbt émet.

- sql.py : tous les `col_cast_double` retirés (fonction supprimée). F15
  (prix/quantite/montant) et puissance_souscrite_kva étaient déjà DOUBLE côté dbt
  → no-op. Les index C15 avant/après et R64 (bigint) cessent d'être castés en DOUBLE.
- Historique : index avant/après Float64 → Int64 (kWh entiers, hérité du bigint dbt).
- Pas de transform Wh→kWh mort à retirer (déjà descendu en dbt à ADR-0034).

Le test de parité (#291) reste le garde-fou de frontière. Empreinte /releves stable
(le mart est inchangé). Fixtures aval + snapshot de structure alignés sur Int64.

+++ electricore/core/models/historique.py
@@ -57,13 +57,15 @@ class Historique(pa.DataFrameModel):
     avant_nature_index: pl.Utf8 | None = pa.Field(nullable=True)
     avant_id_calendrier_fournisseur: pl.Utf8 | None = pa.Field(nullable=True)
     avant_id_calendrier_distributeur: pl.Utf8 | None = pa.Field(nullable=True)
-    avant_index_hp_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    avant_index_hc_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    avant_index_hch_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    avant_index_hph_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    avant_index_hpb_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    avant_index_hcb_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    avant_index_base_kwh: pl.Float64 | None = pa.Field(nullable=True)
+    # Index avant/après en kWh entiers (Int64) : flux_c15 émet du bigint, le loader ne
+    # re-caste plus (ADR-0035) — il trust dbt.
+    avant_index_hp_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    avant_index_hc_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    avant_index_hch_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    avant_index_hph_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    avant_index_hpb_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    avant_index_hcb_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    avant_index_base_kwh: pl.Int64 | None = pa.Field(nullable=True)
 
     # Colonnes de relevés "Après"
     apres_date_releve: DateTime | None = pa.Field(
@@ -72,13 +74,13 @@ class Historique(pa.DataFrameModel):
     apres_nature_index: pl.Utf8 | None = pa.Field(nullable=True)
     apres_id_calendrier_fournisseur: pl.Utf8 | None = pa.Field(nullable=True)
     apres_id_calendrier_distributeur: pl.Utf8 | None = pa.Field(nullable=True)
-    apres_index_hp_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    apres_index_hc_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    apres_index_hch_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    apres_index_hph_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    apres_index_hpb_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    apres_index_hcb_kwh: pl.Float64 | None = pa.Field(nullable=True)
-    apres_index_base_kwh: pl.Float64 | None = pa.Field(nullable=True)
+    apres_index_hp_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    apres_index_hc_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    apres_index_hch_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    apres_index_hph_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    apres_index_hpb_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    apres_index_hcb_kwh: pl.Int64 | None = pa.Field(nullable=True)
+    apres_index_base_kwh: pl.Int64 | None = pa.Field(nullable=True)
 
     # Colonnes ajoutées par l'enrichissement (detecter_points_de_rupture)
     avant_puissance_souscrite: pl.Float64 | None = pa.Field(nullable=True)
